// data/github/current/ui/packages/copilot-for-business/types.ts
export type MenuItem = {
  id: string
  selected: boolean
  title: string
  description: string
  value: string
}

type CopilotForBusinessPolicy = {
  manages: string
  configurable: boolean
  options: MenuItem[]
  visible: boolean | null
}

export type CopilotForBusinessGeneralPolicy = {
  manages: string
  configurable: boolean
  options: MenuItem[]
  visible: boolean | null
  helpurl: string
  helptext: string
  displayname: string
  description: string
  preview: boolean
}

type BingGitHubChatPolicy = {
  manages: string
  configurable: boolean
  enabled: boolean
  options: MenuItem[]
  visible: boolean | null
}

type UserFeedbackPolicy = {
  manages: string
  configurable: boolean
  visible: boolean | null
  enabled: boolean
}

type BetaFeaturesPolicy = {
  manages: string
  configurable: boolean
  visible: boolean | null
  enabled: boolean
}

export type CopilotForBusinessPoliciesPayload = {
  org_name: string
  copilot_plan: string
  editor_chat: CopilotForBusinessPolicy
  mobile_chat: CopilotForBusinessPolicy
  snippy: CopilotForBusinessPolicy
  cli: CopilotForBusinessPolicy
  desktop: CopilotForBusinessPolicy
  copilot_for_dotcom: CopilotForBusinessPolicy
  bing_github_chat: BingGitHubChatPolicy
  copilot_user_feedback_opt_in: UserFeedbackPolicy
  copilot_beta_features_opt_in: BetaFeaturesPolicy
  copilot_extensions: CopilotForBusinessPolicy
  private_telemetry: CopilotForBusinessPolicy
  copilot_usage_metrics_policy: CopilotForBusinessPolicy
  editor_preview_features: CopilotForBusinessPolicy
  a_chat: CopilotForBusinessPolicy
  a_chat_ga: boolean
  a_f: CopilotForBusinessPolicy | CopilotForBusinessGeneralPolicy
  afos: CopilotForBusinessGeneralPolicy
  al: CopilotForBusinessGeneralPolicy
  a_f_ga: boolean
  g_chat: CopilotForBusinessPolicy
  g_chat_ga: boolean
  g_tf: CopilotForBusinessGeneralPolicy
  gtff: CopilotForBusinessGeneralPolicy
  o1: CopilotForBusinessPolicy
  o3: CopilotForBusinessPolicy
  o3_mini_ga: boolean
  o_ff: CopilotForBusinessPolicy
  o_fm: CopilotForBusinessGeneralPolicy
  o_f: CopilotForBusinessPolicy
  o_t: CopilotForBusinessGeneralPolicy
  ofo: CopilotForBusinessGeneralPolicy
  mcp: CopilotForBusinessGeneralPolicy
  overages: CopilotForBusinessPolicy
  swe_agent: CopilotForBusinessGeneralPolicy
  spark: CopilotForBusinessGeneralPolicy
  enterprise_name: string | null
  enterprise_slug: string | null
  docsUrls: {
    generalPrivacyStatement: string
  }
}